refactoring sections: using discriminated unions with one form and one view component per section, renaming people-list to people, removing accordion- and list-section, consistently using two-way binding [()] and model on all form primitives and removing the output signals and onFieldChange().  consistently applying vestFormsViewProviders on all vest form components. Replacing some pipes with computed (e.g. categoryName, svgIcon), repairing maskito mask (maskPredicate was wrongly set on HtmlInputelement instead of HTMLIonInputElement, which resulted in reversed entries.

+++ libs/user/feature/src/lib/user-list.store.ts
@@ -80,13 +80,14 @@ export const UserListStore = signalStore(
         console.log('UserListStore.add() is not yet implemented.');
       },
 
-      async edit(user: UserModel): Promise<void> {
+      async edit(user: UserModel, readOnly = true): Promise<void> {
         store.appNavigationService.pushLink('/user/all' );
-        await navigateByUrl(store.router, `/user/${user.bkey}`);
+        await navigateByUrl(store.router, `/user/${user.bkey}`, { readOnly } );
         store.userResource.reload();
       },
 
-      async delete(user: UserModel): Promise<void> {
+      async delete(user: UserModel, readOnly = true): Promise<void> {
+        if (readOnly) return;
         await store.userService.delete(user, store.appStore.currentUser());
         this.reset();
       },
